requirements.txt
langchain==0.3.25
langchain_community==0.3.23
langchain_mistralai==0.2.10
python-dotenv==1.1.0
